move: App 폴더 생성 후 App 컴포넌트 파일들 이동
