+++ src/Main.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import App from "./App";
+import App from "./components/App";
 import SharedPage from "./pages/SharedPage";
 import FolderPage from "./pages/FolderPage";
 import "./styles/global.css";
