+++ src/components/App/index.js
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
-import Nav from "./components/Nav";
-import Footer from "./components/Footer";
-import "./App.css";
+import Nav from "../Nav";
+import Footer from "../Footer";
+import "./style.css";
 
 const App = ({ user }) => {
   return (
